+++ lib/fastly/models/logging_gcs_response.rb
@@ -55,7 +55,6 @@ module Fastly
     # The name of the GCS bucket.
     attr_accessor :bucket_name
 
-    # The path to upload logs to.
     attr_accessor :path
 
     # A PGP public key that Fastly will use to encrypt your log files before writing them to disk.
@@ -265,6 +264,8 @@ module Fastly
 
       if attributes.key?(:'path')
         self.path = attributes[:'path']
+      else
+        self.path = '/'
       end
 
       if attributes.key?(:'public_key')